+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add the `Media#loadIconFont()` helper method
 * Make the `image` route serve up filetype thumbnails for non-images
 * Add file preview image to the chimera edit view of MediaFile
+* Add thumbnail column to MediaFile chimera index view
 
 ## 0.6.3 (2022-07-23)